FIX: Bump min testing version to 4.5 to match composer requirement (#2)
* FIX: Bump min testing version to 4.5 to match composer requirement

This is necessary due to the updated GraphQL requirement.

Also dropped recipe-testing as we don’t do any Behat tests.

* FIX: Fix broken test.

* FIX: No prefer-source

This shouldn’t be necessary or beneficial and it just slows down the
build.

* FIX: Add gitignore

Helps when running composer install against the module directly

* FIX: Inlined DataObjectFake into this test.

The fakes provided by GraphQL’s test suite are *not* part of its public
API. Given the relative simplicity of this class and the likelihood
that future improvements to it are more likely to break cms-event’s
test suite than enhance it, I’ve opted to copy/paste this fake into this
module rather than expose the Fake as part of GraphQL’s public API.